Fixed: Remove netcls.go
